requirements.txt
CoffeeScript==1.0.10
Django==1.7.2
PyExecJS==1.1.0
django-allauth==0.19.0
django-haystack==2.3.1
django-markdown-deux==1.0.5
django-postman==3.2.0
django-static-precompiler==0.8
djangorestframework==3.0.2
djangular==0.2.7
elasticsearch==1.3.0
markdown2==2.3.0
oauthlib==0.7.2
python-dateutil==2.3
python-openid==2.2.5
pytz==2014.10
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
urllib3==1.10
wsgiref==0.1.2
